restructure api/v1alpha1/condition_test
